product images add

+++ app/forms/product_form.py
@@ -1,8 +1,8 @@
 from flask_wtf import FlaskForm
-from wtforms import StringField, IntegerField, TextAreaField
+from wtforms import StringField, FloatField, TextAreaField
 from wtforms.validators import DataRequired, ValidationError
 
 class ProductForm(FlaskForm):
     name= StringField('name', validators=[DataRequired()])
-    price = IntegerField('price', validators=[DataRequired()])
+    price = FloatField('price', validators=[DataRequired()])
     description = TextAreaField('description', validators=[DataRequired()])
